Fix minor issuses
assignment problems 먼저 생성한 문제부터 출력
끝난 assignment에 문제 추가 불가능하도록 수정
public problem에서 assignment problem으로 가져올때 total_score 같이 저장
public problem에서 assignment problem으로 가져올때 rul_type ASSIGNMENT로 수정
assignment problem에서 다른 user submission 볼 수 없도록 수정
submission이 존재하는 assignment 삭제 불가능하도록 수정

+++ backend/assignment/views/professor.py
@@ -170,6 +170,9 @@ class AssignmentAPI(APIView):
         except Assignment.DoesNotExist:
             return self.error("Assignment does not exist")
 
+        if Submission.objects.filter(assignment_id=assignment_id).exists():
+            return self.error("Can't delete the assignment as it has submissions")
+
         assignment.delete()
         return self.success()
